delete request confirmation and soft delete

+++ app/assets/javascripts/requests.js
@@ -1,30 +1,35 @@
 $(document).ready(function() {
 
-  $('.request-del').click(function(e){
-    var tr = $(this).parents('tr');
-    var url = tr.find('.destroy-url').val();
-    $.ajax({
-      url: url,
-      type: 'DELETE',
-      success: function(resp) {
-        $(tr).remove();
-      }
-    });
-    return false;
-  });
+    $('.request-del').click(function(e){
 
-  $('.request-approve').click(function(e){
-    var tr = $(this).parents('tr');
-    var url = tr.find('.destroy-url').val();
-    $.ajax({
-      url: url,
-      type: 'DELETE',
-      data: { approved: true },
-      success: function(resp) {
-        $(tr).remove();
-      }
+        var tr = $(this).parents('tr');
+        var name = tr.find('.user-name').html();
+        var url = tr.find('.destroy-url').val();
+        if (!confirm("Are you sure you want to reject the enrol request by "+name+"?")) {
+            return
+        }
+        $.ajax({
+            url: url,
+            type: 'DELETE',
+            success: function(resp) {
+                $(tr).remove();
+            }
+        });
+        return false;
     });
-    return false;
-  });
+
+    $('.request-approve').click(function(e){
+        var tr = $(this).parents('tr');
+        var url = tr.find('.destroy-url').val();
+        $.ajax({
+            url: url,
+            type: 'DELETE',
+            data: { approved: true },
+            success: function(resp) {
+                $(tr).remove();
+            }
+        });
+        return false;
+    });
 
 });
